48159 feat(notifications): show completed workflows
